requirements.txt
Flask
Flask-Cors
google-generativeai
Werkzeug
gunicorn